add domains module parsers
