+++ synthaser/parsers.py
@@ -203,6 +203,127 @@ def add_getdb_subparser(subparsers):
     )
 
 
+def add_rules_subparser(subparsers):
+    """Domain rule file generator module"""
+    p = subparsers.add_parser(
+        "rulegen",
+        help="Generate domain rule files",
+    )
+    p.add_argument(
+        "skeleton",
+        type=argparse.FileType(),
+        help="Rules skeleton JSON file",
+    )
+
+
+def add_domain_file_argument(parser):
+    parser.add_argument(
+        "domain_file",
+        type=argparse.FileType("r"),
+        help="CDD family file generated by rulegen parse",
+    )
+
+
+def add_domain_rule_argument(parser):
+    parser.add_argument(
+        "rule_file",
+        help="synthaser Domain rule file"
+    )
+
+
+def add_convert_subparser(subparsers):
+    text = "Convert CDD family names to accessions and vice versa."
+    parser = subparsers.add_parser(
+        "convert",
+        help=text,
+        description=text,
+        epilog="Cameron L.M. Gilchrist, 2020."
+    )
+    add_domain_file_argument(parser)
+    inputs = parser.add_mutually_exclusive_group()
+    inputs.add_argument(
+        "-n",
+        "--names",
+        nargs="+",
+        help="Domain family names"
+    )
+    inputs.add_argument(
+        "-a",
+        "--accessions",
+        nargs="+",
+        help="Domain family accessions"
+    )
+
+
+def add_update_domain_subparser(subparsers):
+    text = "Create domain rule files, add new rules or update existing rules."
+    parser = subparsers.add_parser("update", help=text, description=text)
+    add_domain_file_argument(parser)
+    add_domain_rule_argument(parser)
+
+    single = parser.add_argument_group("Add single rule")
+    single.add_argument("--type", help="Domain type")
+    single.add_argument("--name", help="Domain name")
+    single.add_argument("--families", help="CDD family names", nargs="+")
+
+    multi = parser.add_argument_group("Add multiple rules")
+    multi.add_argument("--file", type=argparse.FileType("r"), help="Skeleton file")
+
+
+def add_remove_rule_subparser(subparsers):
+    parser = subparsers.add_parser(
+        "remove",
+        help="Remove domain rules, or specific CDD families from rules.",
+        description="Remove domain rules, or specific CDD families from rules."
+        " Either remove multiple entire rules, or specify a single rule and multiple"
+        " CDD family accessions using --families."
+    )
+    add_domain_rule_argument(parser)
+    parser.add_argument("--rules", default=[], nargs="+", help="Rule classes")
+    parser.add_argument("--families", default=[], nargs="+", help="CDD family accessions")
+
+
+def add_download_cdd_subparser(subparsers):
+    text = "Download CDD family information from the NCBI."
+    parser = subparsers.add_parser(
+        "download",
+        help=text,
+        description=text,
+    )
+    parser.add_argument("domain_file", help="Output CDD family file")
+    parser.add_argument("--folder", help="Output folder")
+    parser.add_argument("--indent", default=2, help="JSON indent level (def. 2)")
+
+
+def add_summary_subparser(subparsers):
+    text = "Print out domain rules."
+    parser = subparsers.add_parser(
+        "summary",
+        help=text,
+        description=text + " If no rules or families are specified, this module"
+        " will print a summary of the current rules.",
+    )
+    add_domain_rule_argument(parser)
+    parser.add_argument("--rules", default=[], nargs="+", help="Rule classes")
+    parser.add_argument("--families", default=[], nargs="+", help="CDD family accessions")
+
+
+def add_domain_rules_subparser(subparsers):
+    text = "Create or manipulate domain rule files"
+    parser = subparsers.add_parser(
+        "domains",
+        help=text,
+        description=text,
+        epilog="Cameron L.M. Gilchrist, 2020.",
+    )
+    subparsers = parser.add_subparsers(dest="subcommand")
+    add_download_cdd_subparser(subparsers)
+    add_update_domain_subparser(subparsers)
+    add_remove_rule_subparser(subparsers)
+    add_convert_subparser(subparsers)
+    add_summary_subparser(subparsers)
+
+
 def get_parser():
     parser = argparse.ArgumentParser(
         "synthaser",
@@ -213,10 +334,11 @@ def get_parser():
     parser.add_argument(
         "--version", action="version", version="%(prog)s " + __version__
     )
-    subparsers = parser.add_subparsers(dest="subcommand")
+    subparsers = parser.add_subparsers(dest="command")
+    add_search_subparser(subparsers)
+    add_domain_rules_subparser(subparsers)
     add_getdb_subparser(subparsers)
     add_getseq_subparser(subparsers)
-    add_search_subparser(subparsers)
     return parser
 
 
@@ -224,18 +346,18 @@ def parse_args(args):
     parser = get_parser()
     args = parser.parse_args(args)
 
-    if not args.subcommand:
+    if not args.command:
         parser.print_help()
         parser.exit(1)
 
     if (
-        args.subcommand == "search"
+        args.command == "search"
         and args.mode == "remote"
         and args.database not in (None, "cdd", "pfam", "smart", "tigrfam", "cog", "kog")
     ):
         raise ValueError("Expected 'cdd', 'pfam', 'smart', 'tigrfam', 'cog' or 'kog'")
 
-    if args.subcommand == "search" and not any(
+    if args.command == "search" and not any(
         [args.query_ids, args.query_file, args.json_file]
     ):
         raise ValueError("Expected query_ids, query_file or json_file")
